feat(providers): add centralized Providers component with devtools
Move Providers implementation into a dedicated providers module and
augment it with TanStack and React Query devtools, plus a ReleaseEvent
context provider. Initialize a shared QueryClient and keep the Sonner
Toaster for global notifications.

- create providers/index.tsx exporting Providers component that wraps
  children with QueryClientProvider ReleaseEventProvider
- include TanStackDevtools configured to render ReactQueryDevtoolsPanel
  for easier query inspection during development
- import Toaster from ui components to preserve global toasts
- update main.tsx to import from the new providers path and fix JSX
  formatting
- remove the old Providers.tsx file

This consolidates app-level providers, enables better debugging of
React Query state, and clears up module structure.
feat(ui): addEventProvider to ProvidersWrap application children with a ReleaseEventProvider to expose

-related context to the component tree. existing
children are still rendered and retain Toaster and devtools.

This integrates release event state/logic into the app so components
can subscribe via the provider.
tmp
