--- cat-launcher/src/Providers.tsx
@@ -1,18 +1,0 @@
-import { ReactNode } from "react";
-import { QueryClientProvider, QueryClient } from "@tanstack/react-query";
-import { Toaster } from "@/components/ui/sonner";
-
-export interface ProvidersProps {
-  children: ReactNode;
-}
-
-const queryClient = new QueryClient();
-
-export default function Providers({ children }: ProvidersProps) {
-  return (
-    <QueryClientProvider client={queryClient}>
-      {children}
-      <Toaster />
-    </QueryClientProvider>
-  );
-}
