+++ cat-launcher/src/main.tsx
@@ -1,13 +1,13 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "@/App";
-import Providers from "@/Providers";
 import "@/styles/global.css";
+import Providers from "@/providers";
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
     <Providers>
       <App />
     </Providers>
-  </React.StrictMode>,
+  </React.StrictMode>
 );
