+++ cat-launcher/src/providers/index.tsx
@@ -1,0 +1,31 @@
+import { TanStackDevtools } from "@tanstack/react-devtools";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { ReactQueryDevtoolsPanel } from "@tanstack/react-query-devtools";
+import { ReactNode } from "react";
+
+import { Toaster } from "@/components/ui/sonner";
+
+export interface ProvidersProps {
+  children: ReactNode;
+}
+
+const queryClient = new QueryClient();
+
+export default function Providers({ children }: ProvidersProps) {
+  return (
+    <QueryClientProvider client={queryClient}>
+      {children}
+      <Toaster />
+      {import.meta.env.DEV && (
+        <TanStackDevtools
+          plugins={[
+            {
+              name: "TanStack Query",
+              render: <ReactQueryDevtoolsPanel />,
+            },
+          ]}
+        />
+      )}
+    </QueryClientProvider>
+  );
+}
